repository-module/src/main/java/com/example/sonar/multimodule/repository/ProdutoRepository.java
package com.example.sonar.multimodule.repository;

import com.example.sonar.multimodule.entity.Produto;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;

import java.util.List;
import java.util.Optional;

@Component
@RequiredArgsConstructor
public class ProdutoRepository {

    private final JpaProdutoRepository repository;

    public Produto save(Produto produto) {
        return repository.save(produto);
    }

    public List<Produto> findAll() {
        return repository.findAll();
    }

    public Optional<Produto> findById(Long id) {
        return repository.findById(id);
    }

    public void deleteById(Long id) {
        repository.deleteById(id);
    }

}
